Due to python 2 is no longer support, move python 3 version to main package file.
